1.19.80

+++ src/MiNET/MiNET/Net/RakNet/ICustomPacketFactory.cs
@@ -34,6 +34,6 @@ namespace MiNET.Net.RakNet
 	/// </summary>
 	public interface ICustomPacketFactory
 	{
-		public Packet Create(byte messageId, ReadOnlyMemory<byte> buffer, string ns);
+		public Packet Create(short messageId, ReadOnlyMemory<byte> buffer, string ns);
 	}
 }
